Break up test helpers for better code organization and fewer length-related lint failures.
