+++ Tests/SwiftLintFrameworkTests/CompilerProtocolInitRuleTests.swift
@@ -9,8 +9,8 @@ class CompilerProtocolInitRuleTests: XCTestCase {
     }
 
     func testViolationMessageForExpressibleByIntegerLiteral() throws {
-        let config = try XCTUnwrap(makeConfig(nil, ruleID))
-        let allViolations = violations(Example("let a = NSNumber(integerLiteral: 1)"), config: config)
+        let config = try XCTUnwrap(TestHelpers.makeConfig(nil, ruleID))
+        let allViolations = TestHelpers.violations(Example("let a = NSNumber(integerLiteral: 1)"), config: config)
 
         let compilerProtocolInitViolation = allViolations.first { $0.ruleDescription.identifier == ruleID }
         let violation = try XCTUnwrap(
