+++ Tests/SwiftLintFrameworkTests/DeploymentTargetRuleTests.swift
@@ -29,10 +29,10 @@ class DeploymentTargetRuleTests: XCTestCase {
     }
 
     private func violations(_ example: Example, config: Any?) -> [StyleViolation] {
-        guard let config = makeConfig(config, DeploymentTargetRule.description.identifier) else {
+        guard let config = TestHelpers.makeConfig(config, DeploymentTargetRule.description.identifier) else {
             return []
         }
 
-        return SwiftLintFrameworkTests.violations(example, config: config)
+        return TestHelpers.violations(example, config: config)
     }
 }
