+++ Tests/SwiftLintFrameworkTests/FunctionBodyLengthRuleTests.swift
@@ -72,7 +72,7 @@ class FunctionBodyLengthRuleTests: XCTestCase {
     }
 
     private func violations(_ example: Example) -> [StyleViolation] {
-        let config = makeConfig(nil, FunctionBodyLengthRule.description.identifier)!
-        return SwiftLintFrameworkTests.violations(example, config: config)
+        let config = TestHelpers.makeConfig(nil, FunctionBodyLengthRule.description.identifier)!
+        return TestHelpers.violations(example, config: config)
     }
 }
