+++ Tests/SwiftLintFrameworkTests/ModifierOrderTests.swift
@@ -396,12 +396,12 @@ class ModifierOrderTests: XCTestCase {
         }
 
         let ruleID = ModifierOrderRule.description.identifier
-        guard let config = makeConfig(["preferred_modifier_order": ["acl", "final"]], ruleID) else {
+        guard let config = TestHelpers.makeConfig(["preferred_modifier_order": ["acl", "final"]], ruleID) else {
             XCTFail("Failed to create configuration")
             return
         }
 
-        let allViolations = violations(Example("final public var foo: String"), config: config)
+        let allViolations = TestHelpers.violations(Example("final public var foo: String"), config: config)
         let modifierOrderRuleViolation = allViolations.first { $0.ruleDescription.identifier == ruleID }
         if let violation = modifierOrderRuleViolation {
             XCTAssertEqual(violation.reason, "public modifier should be before final.")
