+++ Tests/SwiftLintFrameworkTests/Test Helpers/Example+Testing.swift
@@ -1,0 +1,11 @@
+import SwiftLintFramework
+
+extension Example {
+    var addingEmoji: Example {
+        return with(code: "/* 👨‍👩‍👧‍👦👨‍👩‍👧‍👦👨‍👩‍👧‍👦 */\n\(code)")
+    }
+
+    var addingShebang: Example {
+        return with(code: "#!/usr/bin/env swift\n\(code)")
+    }
+}
