+++ Tests/SwiftLintFrameworkTests/Test Helpers/String+Testing.swift
@@ -1,0 +1,9 @@
+extension String {
+    var formattedAsStringLiteral: String {
+        return "\"" + replacingOccurrences(of: "\n", with: "\\n") + "\""
+    }
+
+    func stringByAppendingPathComponent(_ pathComponent: String) -> String {
+        return bridge().appendingPathComponent(pathComponent)
+    }
+}
