+++ Tests/SwiftLintFrameworkTests/TodoRuleTests.swift
@@ -21,7 +21,7 @@ class TodoRuleTests: XCTestCase {
     }
 
     private func violations(_ example: Example) -> [StyleViolation] {
-        let config = makeConfig(nil, TodoRule.description.identifier)!
-        return SwiftLintFrameworkTests.violations(example, config: config)
+        let config = TestHelpers.makeConfig(nil, TodoRule.description.identifier)!
+        return TestHelpers.violations(example, config: config)
     }
 }
