+++ Tests/SwiftLintFrameworkTests/VerticalWhitespaceRuleTests.swift
@@ -34,11 +34,11 @@ class VerticalWhitespaceRuleTests: XCTestCase {
     }
 
     func testViolationMessageWithMaxEmptyLines() {
-        guard let config = makeConfig(["max_empty_lines": 2], ruleID) else {
+        guard let config = TestHelpers.makeConfig(["max_empty_lines": 2], ruleID) else {
             XCTFail("Failed to create configuration")
             return
         }
-        let allViolations = violations(Example("let aaaa = 0\n\n\n\nlet bbb = 2\n"), config: config)
+        let allViolations = TestHelpers.violations(Example("let aaaa = 0\n\n\n\nlet bbb = 2\n"), config: config)
 
         let verticalWhiteSpaceViolation = allViolations.first { $0.ruleDescription.identifier == ruleID }
         if let violation = verticalWhiteSpaceViolation {
@@ -49,7 +49,7 @@ class VerticalWhitespaceRuleTests: XCTestCase {
     }
 
     func testViolationMessageWithDefaultConfiguration() {
-        let allViolations = violations(Example("let aaaa = 0\n\n\n\nlet bbb = 2\n"))
+        let allViolations = TestHelpers.violations(Example("let aaaa = 0\n\n\n\nlet bbb = 2\n"))
         let verticalWhiteSpaceViolation = allViolations.first(where: { $0.ruleDescription.identifier == ruleID })
         if let violation = verticalWhiteSpaceViolation {
             XCTAssertEqual(violation.reason, "Limit vertical whitespace to a single empty line. Currently 3.")
