+++ Tests/SwiftLintFrameworkTests/XCTSpecificMatcherRuleTests.swift
@@ -122,7 +122,7 @@ class XCTSpecificMatcherRuleTests: XCTestCase {
     }
 
     private func violations(_ example: Example) -> [StyleViolation] {
-        guard let config = makeConfig(nil, XCTSpecificMatcherRule.description.identifier) else { return [] }
-        return SwiftLintFrameworkTests.violations(example, config: config)
+        guard let config = TestHelpers.makeConfig(nil, XCTSpecificMatcherRule.description.identifier) else { return [] }
+        return TestHelpers.violations(example, config: config)
     }
 }
